# api/api.yaml
swagger: "2.0"
info:
  description: "An API to retrieve system data from database"
  version: "1.0.0"
  title: "SRE Data Service"
  contact:
    email: "zayyansmatin@gmail.com"
tags:
- name: "timeseries"
  description: "retrieve timeseries of CPU and concurrency data"
- name: "statistic"
  description: "retrieve average, maximum and minimum of specified resource"
schemes:
- "https"
paths:
  /timeseries:
    get:
      tags:
      - "timeseries"
      produces:
      - "application/json"
      parameters:
      - name: "startTime"
        in: "query"
        required: true
        type: "integer"
      - name: "endTime"
        in: "query"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "string"
        "400":
          description: "Invalid request"
          schema:
            type: "string"
  /statistic/cpu:
    get:
      tags:
      - "statistic"
      summary: "Logs out current logged in user session"
      description: ""
      operationId: "logoutUser"
      produces:
      - "application/json"
      parameters:
      - name: "startTime"
        in: "query"
        required: true
        type: "integer"
      - name: "endTime"
        in: "query"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "string"
        "400":
          description: "Invalid request"
          schema:
            type: "string"
  /statistic/concurrency:
    get:
      tags:
      - "statistic"
      summary: "Logs out current logged in user session"
      produces:
      - "application/json"
      parameters:
      - name: "startTime"
        in: "query"
        required: true
        type: "integer"
      - name: "endTime"
        in: "query"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "string"
        "400":
          description: "Invalid request"
          schema:
            type: "string"
definitions:
  Order:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      petId:
        type: "integer"
        format: "int64"
      quantity:
        type: "integer"
        format: "int32"
      shipDate:
        type: "string"
        format: "date-time"
      status:
        type: "string"
        description: "Order Status"
        enum:
        - "placed"
        - "approved"
        - "delivered"
      complete:
        type: "boolean"
        default: false
    xml:
      name: "Order"
  Category:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Category"
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
      userStatus:
        type: "integer"
        format: "int32"
        description: "User Status"
    xml:
      name: "User"
  Tag:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Tag"
  Pet:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      id:
        type: "integer"
        format: "int64"
      category:
        $ref: "#/definitions/Category"
      name:
        type: "string"
        example: "doggie"
      photoUrls:
        type: "array"
        xml:
          name: "photoUrl"
          wrapped: true
        items:
          type: "string"
      tags:
        type: "array"
        xml:
          name: "tag"
          wrapped: true
        items:
          $ref: "#/definitions/Tag"
      status:
        type: "string"
        description: "pet status in the store"
        enum:
        - "available"
        - "pending"
        - "sold"
    xml:
      name: "Pet"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"